Create the intial style and structure of the website
This is the first commit with actual website code. This commit is very
dense because I did not seperate my initial progress into individual
commits. That was a mistake.

I've added a default layout, "business card" and footer includes,
a style sheet, a Gemfile, the initial homepage, and a placeholder
image.

This was to setup the structure and style of the website. This is not
complete. I still need a nav bar, blog page, and a few extra pages to
organize my content better.
fasdfasdfasdfasdfI added a default
layout, reusable "business card" and footer includes, main stylesheet,